+++ src/main/java/com/shippingflow/presenter/api/item/request/CreateItemRequest.java
@@ -1,4 +1,4 @@
-package com.shippingflow.presenter.api.item.controller.request;
+package com.shippingflow.presenter.api.item.request;
 
 import com.shippingflow.core.usecase.aggregate.item.CreateItemUseCase;
 import jakarta.validation.constraints.NotBlank;
