+++ src/main/java/com/shippingflow/presenter/api/item/request/validator/StockTransactionTypeValidator.java
@@ -1,4 +1,4 @@
-package com.shippingflow.presenter.api.item.controller.request.validator;
+package com.shippingflow.presenter.api.item.request.validator;
 
 import com.shippingflow.core.domain.aggregate.item.model.local.StockTransactionType;
 import io.micrometer.common.util.StringUtils;
