+++ src/main/java/com/shippingflow/presenter/api/item/response/ItemsResponse.java
@@ -1,8 +1,8 @@
-package com.shippingflow.presenter.api.item.controller.response;
+package com.shippingflow.presenter.api.item.response;
 
 import com.shippingflow.core.domain.aggregate.item.dto.ItemWithStockDto;
 import com.shippingflow.core.domain.common.pagination.PageResponse;
-import com.shippingflow.presenter.api.common.response.PageApiResponse;
+import com.shippingflow.presenter.common.response.PageApiResponse;
 
 import java.util.List;
