+++ src/main/java/com/shippingflow/presenter/api/item/response/StockTransactionsResponse.java
@@ -1,9 +1,9 @@
-package com.shippingflow.presenter.api.item.controller.response;
+package com.shippingflow.presenter.api.item.response;
 
 import com.shippingflow.core.domain.aggregate.item.dto.StockTransactionDto;
 import com.shippingflow.core.domain.common.pagination.PageResponse;
 import com.shippingflow.core.usecase.aggregate.item.GetStockTransactionsUseCase;
-import com.shippingflow.presenter.api.common.response.PageApiResponse;
+import com.shippingflow.presenter.common.response.PageApiResponse;
 
 import java.util.List;
