+++ src/main/java/com/shippingflow/presenter/common/response/PageApiResponse.java
@@ -1,4 +1,4 @@
-package com.shippingflow.presenter.api.common.response;
+package com.shippingflow.presenter.common.response;
 
 public record PageApiResponse(int page, int size, long totalElements, int totalPages) {
     public static PageApiResponse of(int page, int size, long totalElements, int totalPages) {
