+++ src/test/java/com/shippingflow/presenter/WebMvcTestSupport.java
@@ -1,7 +1,7 @@
 package com.shippingflow.presenter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.shippingflow.presenter.api.item.controller.ItemController;
+import com.shippingflow.presenter.api.item.ItemController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
